moved exception to wrappers package
