+++ app/src/main/java/gov/hhs/cdc/trustedintermediary/etor/orders/OrderController.java
@@ -1,8 +1,8 @@
 package gov.hhs.cdc.trustedintermediary.etor.orders;
 
 import gov.hhs.cdc.trustedintermediary.domainconnector.DomainRequest;
-import gov.hhs.cdc.trustedintermediary.etor.FhirParseException;
 import gov.hhs.cdc.trustedintermediary.external.hapi.HapiOrder;
+import gov.hhs.cdc.trustedintermediary.wrappers.FhirParseException;
 import gov.hhs.cdc.trustedintermediary.wrappers.HapiFhir;
 import gov.hhs.cdc.trustedintermediary.wrappers.Logger;
 import javax.inject.Inject;
